# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "goose-server",
    "description": "An AI agent",
    "contact": {
      "name": "Block",
      "email": "ai-oss-tools@block.xyz"
    },
    "license": {
      "name": "Apache-2.0"
    },
    "version": "1.1.0"
  },
  "paths": {
...
    "/sessions": {
      "get": {
        "tags": [
          "Session Management"
        ],
        "operationId": "list_sessions",
        "responses": {
          "200": {
            "description": "List of available sessions retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing API key"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/sessions/{session_id}": {
      "get": {
        "tags": [
          "Session Management"
        ],
        "operationId": "get_session_history",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "description": "Unique identifier for the session",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session history retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionHistoryResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing API key"
          },
          "404": {
            "description": "Session not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "put": {
        "tags": [
          "Session Management"
        ],
        "operationId": "update_session",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "description": "Unique identifier for the session",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Session updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateSessionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing API key"
          },
          "404": {
            "description": "Session not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "Annotations": {
        "type": "object",
        "properties": {
          "audience": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Role"
            }
          },
          "priority": {
            "type": "number"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ContextLengthExceeded": {
        "type": "object",
        "required": [
          "msg"
        ],
        "properties": {
          "msg": {
            "type": "string"
          }
        }
      },
...
      "FrontendToolRequest": {
        "type": "object",
        "required": [
          "id",
          "toolCall"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "toolCall": {
            "type": "object"
          }
        }
      },
      "ImageContent": {
        "type": "object",
        "required": [
          "data",
          "mimeType"
        ],
        "properties": {
          "annotations": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Annotations"
              }
            ]
          },
          "data": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          }
        }
      },
...
      "Message": {
        "type": "object",
        "description": "A message to or from an LLM",
        "required": [
          "role",
          "created",
          "content"
        ],
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MessageContent"
            }
          },
          "created": {
            "type": "integer",
            "format": "int64"
          },
          "id": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        }
      },
      "MessageContent": {
        "oneOf": [
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/TextContent"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "text"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/ImageContent"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "image"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/ToolRequest"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "toolRequest"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/ToolResponse"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "toolResponse"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/ToolConfirmationRequest"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "toolConfirmationRequest"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/FrontendToolRequest"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "frontendToolRequest"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/ThinkingContent"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "thinking"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/RedactedThinkingContent"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "redactedThinking"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/ContextLengthExceeded"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "contextLengthExceeded"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/SummarizationRequested"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "summarizationRequested"
                    ]
                  }
                }
              }
            ]
          }
        ],
        "description": "Content passed inside a message, which can be both simple content and tool content",
        "discriminator": {
          "propertyName": "type"
        }
      },
...
      "RedactedThinkingContent": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "string"
          }
        }
      },
...
      "Role": {
        "oneOf": [
          {
            "type": "string"
          },
          {
            "type": "string"
          }
        ]
      },
...
      "SessionHistoryResponse": {
        "type": "object",
        "required": [
          "sessionId",
          "metadata",
          "messages"
        ],
        "properties": {
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Message"
            },
            "description": "List of messages in the session conversation"
          },
          "metadata": {
            "$ref": "#/components/schemas/SessionMetadata"
          },
          "sessionId": {
            "type": "string",
            "description": "Unique identifier for the session"
          }
        }
      },
      "SessionInfo": {
        "type": "object",
        "required": [
          "id",
          "path",
          "modified",
          "metadata"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "metadata": {
            "$ref": "#/components/schemas/SessionMetadata"
          },
          "modified": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        }
      },
      "SessionListResponse": {
        "type": "object",
        "required": [
          "sessions"
        ],
        "properties": {
          "sessions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SessionInfo"
            },
            "description": "List of available session information objects"
          }
        }
      },
      "SessionMetadata": {
        "type": "object",
        "description": "Metadata for a session, stored as the first line in the session file",
        "required": [
          "working_dir",
          "description",
          "message_count"
        ],
        "properties": {
          "accumulated_input_tokens": {
            "type": "integer",
            "format": "int32",
            "description": "The number of input tokens used in the session. Accumulated across all messages.",
            "nullable": true
          },
          "accumulated_output_tokens": {
            "type": "integer",
            "format": "int32",
            "description": "The number of output tokens used in the session. Accumulated across all messages.",
            "nullable": true
          },
          "accumulated_total_tokens": {
            "type": "integer",
            "format": "int32",
            "description": "The total number of tokens used in the session. Accumulated across all messages (useful for tracking cost over an entire session).",
            "nullable": true
          },
          "description": {
            "type": "string",
            "description": "A short description of the session, typically 3 words or less"
          },
          "input_tokens": {
            "type": "integer",
            "format": "int32",
            "description": "The number of input tokens used in the session. Retrieved from the provider's last usage.",
            "nullable": true
          },
          "message_count": {
            "type": "integer",
            "description": "Number of messages in the session",
            "minimum": 0
          },
          "output_tokens": {
            "type": "integer",
            "format": "int32",
            "description": "The number of output tokens used in the session. Retrieved from the provider's last usage.",
            "nullable": true
          },
          "project_id": {
            "type": "string",
            "description": "ID of the project this session belongs to, if any",
            "nullable": true
          },
          "schedule_id": {
            "type": "string",
            "description": "ID of the schedule that triggered this session, if any",
            "nullable": true
          },
          "total_tokens": {
            "type": "integer",
            "format": "int32",
            "description": "The total number of tokens used in the session. Retrieved from the provider's last usage.",
            "nullable": true
          },
          "working_dir": {
            "type": "string",
            "description": "Working directory for the session",
            "example": "/home/user/sessions/session1"
          }
        }
      },
...
      "SummarizationRequested": {
        "type": "object",
        "required": [
          "msg"
        ],
        "properties": {
          "msg": {
            "type": "string"
          }
        }
      },
      "TextContent": {
        "type": "object",
        "required": [
          "text"
        ],
        "properties": {
          "annotations": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Annotations"
              }
            ]
          },
          "text": {
            "type": "string"
          }
        }
      },
      "ThinkingContent": {
        "type": "object",
        "required": [
          "thinking",
          "signature"
        ],
        "properties": {
          "signature": {
            "type": "string"
          },
          "thinking": {
            "type": "string"
          }
        }
      },
...
      "ToolConfirmationRequest": {
        "type": "object",
        "required": [
          "id",
          "toolName",
          "arguments"
        ],
        "properties": {
          "arguments": {},
          "id": {
            "type": "string"
          },
          "prompt": {
            "type": "string",
            "nullable": true
          },
          "toolName": {
            "type": "string"
          }
        }
      },
...
      "ToolRequest": {
        "type": "object",
        "required": [
          "id",
          "toolCall"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "toolCall": {
            "type": "object"
          }
        }
      },
      "ToolResponse": {
        "type": "object",
        "required": [
          "id",
          "toolResult"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "toolResult": {
            "type": "object"
          }
        }
      },
...
      "UpdateSessionRequest": {
        "type": "object",
        "required": [
          "description"
        ],
        "properties": {
          "description": {
            "type": "string",
            "description": "New description for the session"
          }
        }
      },
      "UpdateSessionResponse": {
        "type": "object",
        "required": [
          "success",
          "metadata"
        ],
        "properties": {
          "metadata": {
            "$ref": "#/components/schemas/SessionMetadata"
          },
          "success": {
            "type": "boolean",
            "description": "Success status"
          }
        }
      },
...
    }
  }
}